Demo complete
I now have a basic example of a WAR project that serves all of its pages
via a single angular-powered JSP with real links that are random
accessible.

The key is that the web.xml serves all pages matching /app/* via the
JSP, so long as all of the angular routes begin with <our context
root>/app we should be serving via the app.jsp file.

This also frees us up to serve up REST endpoints via another URL
pattern, such as /api